+++ miniwin/src/internal/meshutils.h
@@ -11,5 +11,5 @@ void FlattenSurfaces(
 	const size_t indexCount,
 	bool hasTexture,
 	std::vector<D3DRMVERTEX>& dedupedVertices,
-	std::vector<DWORD>& newIndices
+	std::vector<uint16_t>& newIndices
 );
